chore(test): move tests to /tests and remove /e2e
