+++ tests/smoke-navigation.spec.ts
@@ -1,4 +1,4 @@
-import { test, expect } from '../fixtures';
+import { test, expect } from './fixtures';
 
 /**
  * Smoke & Navigation Tests
